Updated README file and change SIG digits
Added additional support macro files to display the additional config
file parameters that will be added to the Klipper printer object

Added a HEATERS_STATUS to zero_overshoot extension to make  debuging
issues easier.

Change the number of significant digits from 6 to 3
because that is all  klipper will save to the printer.cfg file when doing a
SAVE_CONFIG command.

Remove `import configparser` It is not needed